+++ tests/TestCase/Exception/UnsupportedImageTypeExceptionTest.php
@@ -14,6 +14,7 @@ declare(strict_types=1);
  */
 namespace Thumber\Test\Exception;
 
+use PHPUnit\Framework\Error\Deprecated;
 use Thumber\Exception\UnsupportedImageTypeException;
 use Thumber\TestSuite\TestCase;
 
@@ -32,14 +33,21 @@ class UnsupportedImageTypeExceptionTest extends TestCase
             throw new UnsupportedImageTypeException();
         } catch (UnsupportedImageTypeException $e) {
             $this->assertSame('Image type not supported by this driver', $e->getMessage());
-            $this->assertNull($e->getImageType());
+            $this->assertNull($e->getValue());
         }
 
         try {
             throw new UnsupportedImageTypeException('', 0, null, 'example');
         } catch (UnsupportedImageTypeException $e) {
             $this->assertSame('Image type `example` is not supported by this driver', $e->getMessage());
-            $this->assertSame('example', $e->getImageType());
+            $this->assertSame('example', $e->getValue());
         }
+
+        $current = error_reporting(E_ALL & ~E_USER_DEPRECATED);
+        $e->getImageType();
+        error_reporting($current);
+
+        $this->expectException(Deprecated::class);
+        $e->getImageType();
     }
 }